endlessos: Explicitly use UTF-8 as translation charset
If the character set isn't specified, then you'll get whatever is the
libc default. On glibc that's ISO-8859-1 for en_US. Not only is that not
UTF-8 like we want, but it's increasingly likely that the locale isn't
installed. In that case get_language_from_locale returns NULL and we use
the bare locale. Just use UTF-8 explicitly.

Signed-off-by: Dan Nicholson <dbn@endlessos.org>
Signed-off-by: Will Thompson <wjt@endlessos.org>