tests: fix parameterization of objects
The parameterization process needs to assign a unique ID to every test
instance. For most object types this is done by looking for the
'filename' attribute, falling back to str(obj) if the attr does not
exist. None of the object types have a 'filename' attribute, however,
it should have been 'path'.

This would have not been fatal if the previous bug of lacking __repr__
method on Device hadn't also been present. These combined to result
in a ID that reflected the python object memory address which is
unstable across execution attempts.

Fixing this ensures that pytest-xdist can process our tests in
parallel.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>