Makefile: support running pytest in parallel
With this change, 'make check' will make use of xdist to run tests in
parallel when both:

 * pytest xdist plugin is installed
 * parallel make is requested

IOW,  'make -j 20 check' will run tests in parallel but still a plain
'make check' will not. The latter is more helpful for debugging since
the output shows details of which test is being run.

Closes: https://gitlab.com/libosinfo/osinfo-db/-/issues/140
Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>