Provide a correct short-id for some Ubuntu releases
It looks like the ubuntu-14.04.xml.in (Ubuntu Trusty) file was copied and when
the short-id for the code name was changed it was accidentally truncated to
"ubuntut" instead of "ubuntu" then the code name for that file was appended.

This has resulted in:

14.10 -> ubuntututopic
15.04 -> ubuntuvivid
15.10 -> ubuntutwily
16.04 -> ubuntutxenial
16.10 -> ubuntutyakkety
17.04 -> ubuntutzesty

This commit provides the correct short-ids without that wayward "t".

I have left the bogus short-ids in place for people (like me) who have applied
fixes to handle these who might otherwise find their own code broken.

Signed-off-by: Andrew Ruthven <puck@catalyst.net.nz>